ai-testing-resource/requirements.txt
# Core
flask>=3.0.0
python-dotenv>=1.0.0

# AI/ML
anthropic>=0.39.0
tiktoken>=0.5.0

# ChromaDB and dependencies - pinned for NumPy compatibility
# Note: ChromaDB 0.5.23 requires numpy<2 due to onnxruntime
numpy>=1.22.5,<2.0.0
chromadb>=0.5.23,<0.6.0
sentence-transformers>=2.2.0

# Syntax highlighting
pygments>=2.16.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Phase 2: TSR and Monitoring
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# WebSocket support
flask-socketio>=5.3.0
python-socketio>=5.10.0

# PDF generation (optional - for TSR export)
# reportlab>=4.0.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
